pyproject.toml
[project]
name = "butterfly-classification"
version = "0.1.0"
description = ""
authors = [
    {name = "Sardor Lutfullaev",email = "lutfullaevs03@gmail.com"}
]
readme = "README.md"
[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
dvc = "^3.50.0"
pre-commit = ">=4.2.0,<5.0.0"
dvc-s3 = "^3.1.0"
numpy = "<2.0"
ipykernel = "^6.29.5"
Pillow = "^10.0.0"
scikit-learn = "^1.3.0"
seaborn = "^0.12.0"
pytorch-lightning = "^2.4.0"
torch = "^2.5.1"
torchvision = "^0.20.1"
torchmetrics = "^1.5.2"
mlflow = "^2.19.0"
dvc-gdrive = "^3.0.1"
pyopenssl = ">=23.0.0"
kagglehub = "^0.3.12"

[tool.poetry.group.dev.dependencies]
pandas = "^2.2.3"
lightning = "^2.1.3"
matplotlib = "^3.10.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"







[tool.poetry.group.conversion]
optional = true

[tool.poetry.group.conversion.dependencies]
# Базовые зависимости для ONNX (поддерживаются на всех платформах)
onnx = "^1.12.0"
onnxruntime = "^1.12.0"
onnxruntime-gpu = {version = "^1.12.0", optional = true}

# TensorRT зависимости (только для Linux с NVIDIA GPU)
# Примечание: TensorRT не поддерживается на macOS и требует ручной установки NVIDIA TensorRT SDK
tensorrt = {version = "^8.4.0", optional = true, markers = "platform_system == 'Linux'"}
pycuda = {version = "^2022.1", optional = true, markers = "platform_system == 'Linux'"}





[tool.poetry.group.gpu]
optional = true

[tool.poetry.group.gpu.dependencies]
torch = ">2.4.0"
torchvision = "^0.20.0"
mlflow = "^2.19.0"






[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
pylint = "^3.2.5"





[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
coverage = "^7.5.4"
pytest = "^8.0.2"
pytest-cov = "^5.0.0"
